Fix markup and naming
