+++ app/assets/javascripts/sitewide/file_preview.js
@@ -449,7 +449,7 @@ var FilePreviewModal = (function() {
             if (!readOnly && data.editable) {
               modal.find('.file-edit-link').css('display', '');
               modal.find('.file-edit-link').off().click(function(ev) {
-                $.post('/files/' + data.id + '/start_edit');
+                $.post('/files/' + data.id + '/start_edit_image');
                 ev.preventDefault();
                 ev.stopPropagation();
                 modal.modal('hide');
