+++ config/initializers/extends.rb
@@ -233,7 +233,7 @@ class Extends
     experiment: [*27..31, 57],
     reports: [48, 50, 49],
     inventories: [70, 71, 105, 72, 73, 74, 102, 75, 76, 77, 78, 96, 107],
-    protocol_repository: [80, 103, 89, 87, 79, 90, 91, 88, 85, 86, 84, 81, 82, 83, 101, 102],
+    protocol_repository: [80, 103, 89, 87, 79, 90, 91, 88, 85, 86, 84, 81, 82, 83, 101, 112],
     team: [92, 94, 93, 97, 104]
   }.freeze
 end
